requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
tqdm>=4.65.0
wandb>=0.15.0
einops>=0.7.0
h5py>=3.9.0
Pillow>=10.0.0
scikit-learn>=1.3.0
networkx>=3.1
pycocotools>=2.0.7
